+++ python/marvin/db/models/DapModelClasses.py
@@ -398,6 +398,17 @@ if 'testtable' in db.engine.table_names('mangadapdb'):
         def __repr__(self):
             return ('<TestTable pk={0}, file={1}>'.format(self.pk, self.file_pk))
 
+if 'cleanspaxeltest' in db.engine.table_names('mangadapdb'):
+    class CleanSpaxelTest(Base):
+        __tablename__ = 'cleanspaxeltest'
+        __table_args__ = {'autoload': True, 'schema': 'mangadapdb'}
+
+        pk = Column(Integer, primary_key=True)
+        file = relationship(File, backref="cleanspaxeltest")
+
+        def __repr__(self):
+            return ('<CleanSpaxelTest pk={0}, file={1}, val={2}>'.format(self.pk, self.file_pk, self.value))
+
 # -----
 # Buld Relationships
 # -----
@@ -454,6 +465,7 @@ except RuntimeError as error:
     print('DapModelClasses, Error during configure_mapper: {0}'.format(error))
 
 dap_cache = RelationshipCache(File.cube).\
+               and_(RelationshipCache(File.pipelineinfo)).\
                and_(RelationshipCache(File.filetype)).\
                and_(RelationshipCache(File.structure)).\
                and_(RelationshipCache(ModelCube.file)).\