+++ python/marvin/web/extensions.py
@@ -9,6 +9,7 @@
 # @Last Modified time: 2018-06-05 11:39:50
 
 from __future__ import print_function, division, absolute_import
+from flask_compress import Compress
 from flask_featureflags import FeatureFlag
 from raven.contrib.flask import Sentry
 from flask_limiter import Limiter
@@ -18,6 +19,7 @@ from flask_caching import Cache
 from flask_login import LoginManager
 from flask_jwt_extended import JWTManager
 from flask_cors import CORS
+from flask_session import Session
 import flask_jsglue as jsg
 import logging
 
@@ -47,3 +49,9 @@ jwt = JWTManager()
 
 # Flask CORS
 cors = CORS()
+
+# Flask Session
+session = Session()
+
+# Flask Compress
+compress = Compress()